Add two script to backup geth data and chain data
the two script are added to backup geth data and wan-chain data at the first period of wan-chain. 
（Use admin.exportChain and comress the data pakage）
And the backup frequency is every 3 hours, a week of data would be backuped.
Later a set_env script would be added to see the environment vaiable, and crontab would be also auto added.